requirements.txt
graphviz
terminaltables
pint
plotly
networkx
CIRpy
colorama
pyserial
pyyaml
flask
aiohttp
cython
vedis
schedule
requests
sphinx
numpy
pandas
